RavenDB-18104 - POST /databases/*/subscriptions/try
