+++ src/Raven.Server/Documents/Handlers/SubscriptionsHandler.cs
@@ -1,6 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Diagnostics;
 using System.IO;
 using System.Linq;
 using System.Net;
@@ -8,17 +6,9 @@ using System.Threading.Tasks;
 using Raven.Client;
 using Raven.Client.Documents.Changes;
 using Raven.Client.Documents.Subscriptions;
-using Raven.Client.Exceptions.Documents.Subscriptions;
-using Raven.Client.Extensions;
 using Raven.Server.Documents.Handlers.Processors.Subscriptions;
-using Raven.Server.Documents.Includes;
-using Raven.Server.Documents.Replication;
-using Raven.Server.Documents.Sharding.Handlers.Processors.Subscriptions;
-using Raven.Server.Documents.Sharding.Subscriptions;
 using Raven.Server.Documents.Subscriptions;
-using Raven.Server.Documents.Subscriptions.SubscriptionProcessor;
 using Raven.Server.Documents.TcpHandlers;
-using Raven.Server.Json;
 using Raven.Server.Routing;
 using Raven.Server.ServerWide.Context;
 using Raven.Server.TrafficWatch;
@@ -32,147 +22,8 @@ namespace Raven.Server.Documents.Handlers
         [RavenAction("/databases/*/subscriptions/try", "POST", AuthorizationStatus.ValidUser, EndpointType.Write, DisableOnCpuCreditsExhaustion = true)]
         public async Task Try()
         {
-            using (ContextPool.AllocateOperationContext(out DocumentsOperationContext context))
-            {
-                BlittableJsonReaderObject json = await context.ReadForMemoryAsync(RequestBodyStream(), null);
-                SubscriptionTryout tryout = JsonDeserializationServer.SubscriptionTryout(json);
-
-                SubscriptionConnection.ParsedSubscription sub = SubscriptionConnection.ParseSubscriptionQuery(tryout.Query);
-                SubscriptionPatchDocument patch = null;
-                if (string.IsNullOrEmpty(sub.Script) == false)
-                {
-                    patch = new SubscriptionPatchDocument(sub.Script, sub.Functions);
-                }
-
-                if (sub.Collection == null)
-                    throw new ArgumentException("Collection must be specified");
-
-                const int maxPageSize = 1024;
-                var pageSize = GetIntValueQueryString("pageSize") ?? 1;
-                if (pageSize > maxPageSize)
-                    throw new ArgumentException($"Cannot gather more than {maxPageSize} results during tryouts, but requested number was {pageSize}.");
-
-                var state = new SubscriptionState
-                {
-                    ChangeVectorForNextBatchStartingPoint = tryout.ChangeVector,
-                    Query = tryout.Query
-                };
-
-                
-                var includeCmd = new IncludeDocumentsCommand(Database.DocumentsStorage, context, sub.Includes, isProjection: patch != null);
-
-                if (Enum.TryParse(
-                    tryout.ChangeVector,
-                    out Constants.Documents.SubscriptionChangeVectorSpecialStates changeVectorSpecialValue))
-                {
-                    switch (changeVectorSpecialValue)
-                    {
-                        case Constants.Documents.SubscriptionChangeVectorSpecialStates.BeginningOfTime:
-                        case Constants.Documents.SubscriptionChangeVectorSpecialStates.DoNotChange:
-                            state.ChangeVectorForNextBatchStartingPoint = null;
-                            break;
-
-                        case Constants.Documents.SubscriptionChangeVectorSpecialStates.LastDocument:
-                            using (context.OpenReadTransaction())
-                            {
-                                state.ChangeVectorForNextBatchStartingPoint = Database.DocumentsStorage.GetLastDocumentChangeVector(context.Transaction.InnerTransaction, context, sub.Collection);
-                            }
-                            break;
-                    }
-                }
-                else
-                {
-                    state.ChangeVectorForNextBatchStartingPoint = tryout.ChangeVector;
-                }
-
-                var changeVector = state.ChangeVectorForNextBatchStartingPoint.ToChangeVector();
-                var cv = changeVector.FirstOrDefault(x => x.DbId == Database.DbBase64Id);
-
-                var sp = Stopwatch.StartNew();
-                var timeLimit = TimeSpan.FromSeconds(GetIntValueQueryString("timeLimit", false) ?? 15);
-                var startEtag = cv.Etag;
-
-                SubscriptionProcessor processor;
-                if(sub.Revisions)
-                    processor = new TestRevisionsSubscriptionProcessor(Server.ServerStore, Database, state, sub.Collection, new SubscriptionWorkerOptions("dummy"), new IPEndPoint(HttpContext.Connection.RemoteIpAddress, HttpContext.Connection.RemotePort));
-                else
-                    processor = new TestDocumentsSubscriptionProcessor(Server.ServerStore, Database, state, sub.Collection, new SubscriptionWorkerOptions("dummy"), new IPEndPoint(HttpContext.Connection.RemoteIpAddress, HttpContext.Connection.RemotePort));
-                processor.AddScript(patch);
-
-                using (processor)
-                await using (var writer = new AsyncBlittableJsonTextWriter(context, ResponseBodyStream()))
-                using (Database.ServerStore.Engine.ContextPool.AllocateOperationContext(out ClusterOperationContext clusterOperationContext))
-                using (clusterOperationContext.OpenReadTransaction())
-                using (context.OpenReadTransaction())
-                {
-                    processor.InitializeForNewBatch(clusterOperationContext, context, includeCmd);
-
-                    writer.WriteStartObject();
-                    writer.WritePropertyName("Results");
-                    writer.WriteStartArray();
-                    var numberOfDocs = 0;
-                    while (numberOfDocs == 0 && sp.Elapsed < timeLimit)
-                    {
-                        var first = true;
-                        var lastEtag = startEtag;
-                        
-                        ((IEtagSettable)processor).SetStartEtag(startEtag);
-
-                        foreach (var itemDetails in processor.GetBatch())
-                        {
-                            if (itemDetails.Doc.Data != null)
-                            {
-                                using (itemDetails.Doc.Data)
-                                {
-                                    includeCmd.Gather(itemDetails.Doc);
-
-                                    if (first == false)
-                                        writer.WriteComma();
-
-                                    if (itemDetails.Exception == null)
-                                    {
-                                        writer.WriteDocument(context, itemDetails.Doc, metadataOnly: false);
-                                    }
-                                    else
-                                    {
-                                        var documentWithException = new DocumentWithException
-                                        {
-                                            Exception = itemDetails.Exception.ToString(),
-                                            ChangeVector = itemDetails.Doc.ChangeVector,
-                                            Id = itemDetails.Doc.Id,
-                                            DocumentData = itemDetails.Doc.Data
-                                        };
-                                        writer.WriteObject(context.ReadObject(documentWithException.ToJson(), ""));
-                                    }
-
-                                    first = false;
-
-                                    if (++numberOfDocs >= pageSize)
-                                        break;
-                                }
-                            }
-
-                            if (sp.Elapsed >= timeLimit)
-                                break;
-
-                            lastEtag = itemDetails.Doc.Etag;
-                        }
-
-                        if (startEtag == lastEtag)
-                            break;
-
-                        startEtag = lastEtag;
-                    }
-
-                    writer.WriteEndArray();
-                    writer.WriteComma();
-                    writer.WritePropertyName("Includes");
-                    var includes = new List<Document>();
-                    includeCmd.Fill(includes, includeMissingAsNull: false);
-                    await writer.WriteIncludesAsync(context, includes);
-                    writer.WriteEndObject();
-                }
-            }
+            using (var processor = new SubscriptionsHandlerProcessorForTrySubscription(this))
+                await processor.ExecuteAsync();
         }
 
         [RavenAction("/databases/*/subscriptions", "PUT", AuthorizationStatus.ValidUser, EndpointType.Write)]
